feature: wallet history placeholder details: - Modify my wallet image - wallet history screen place holder

+++ lib/data/models/wallet/wallet_summary_model.dart
@@ -6,7 +6,7 @@ part 'wallet_summary_model.g.dart';
 @freezed
 abstract class WalletSummaryModel with _$WalletSummaryModel {
   const factory WalletSummaryModel({
-    required int totalDiamondBalance,
+    required double totalDiamondBalance,
     required int vexBalance,
     required String monthlyCardStatus, // active | expired | none
   }) = _WalletSummaryModel;
